sdks/python/pyproject.toml
[project]
name = "warrant-sdk"
version = "0.1.0"
description = "Python SDK for the Warrant cryptographic agent credential service"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyjwt[crypto]>=2.8",
]

[project.optional-dependencies]
langgraph = ["langgraph>=0.2", "langchain-core>=0.3"]
dev = ["pytest", "pytest-asyncio", "respx"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["warrant"]
